Added new features and refactoring
- Added saving and loading to and from files
* As of now, only for the player stats and player inventory
* Future plans to add similar functionality for rooms, items, etc
- Changed the way the player gets constructed based off of files
- Added save action so the player can save whenever
- Added 'isGameRunning' flag for main game loop
- Added an error check in the equip function
- Added an error check when moving rooms
- Added getters and setters for room, weapon, and armor indices in the player object for file loading and saving
- Added method in player object to add recipes
* Will be useful later when I add a way to acquire new recipes
- Much refactoring of text and values throughout

+++ src/dev/tesch/Items/Armor.java
@@ -6,7 +6,7 @@ public class Armor extends Item {
             String n,
             String d,
             String uMessage,
-            double itW,
+            int itW,
             boolean cPickup,
             boolean cUse,
             boolean cC,
